Create Calculate Volume of a Circular tank
This code creates two input fields for the tank radius and height and a button to calculate the tank volume. When the user clicks the "Calculate" button, the calculateVolume() function is called. The function retrieves the values from the input fields, calculates the volume of the tank using the formula πr²h, and displays the result on the page using the textContent property of the span element with the id "volume".

Note that in this example, the input fields are in meters and the volume is in cubic meters. Also, keep in mind that this calculation assumes the tank is a perfect cylinder. If the tank has a different shape, the calculation will need to be adjusted accordingly.

I hope this helps!