Create Volume of a cuboid
In this code, we have created a form with three input fields for the length, width, and height of the cuboid. We have also included a button that triggers a JavaScript function called calculateVolume() when clicked. This function retrieves the values of the input fields, calculates the volume of the cuboid using the formula length * width * height, and displays the result in the result div.

We have also included the Bootstrap 5 CSS and JS files in the head and body of the HTML document to style and add functionality to our application.

I hope this helps you create a simple cuboid volume calculator using HTML 5, Bootstrap 5, and JavaScript!